Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

codeworkx put the blob in the vendor repo - this should be ready to go now.

Google Wallet will hose up NFC on any device it is installed on currently.  Also, the device needs to be rebooted once after first installation of this patch to permit the secure element to be wiped.  NFC apps will hang until then.

Patch-set: 5